avoid dup apps.
