+++ Utility/Resource.py
@@ -446,78 +446,3 @@ def getStrRatioSimilarity(s, t, usePartial=False):
     if usePartial:
         return fuzz.partial_ratio(s.lower(), t.lower())
     return fuzz.ratio(s.lower(), t.lower())
-
-
-def getAppDictEntry(app, update=True):
-    entry = None
-    appName = None
-    appPkgName = None
-
-    if type(app) == dict and "application" in app:
-        appName = app["application"]["application_name"]
-        appPkgName = appName + (" (%s)" % app["application"]["package_name"])
-        entry = {
-            "app_name": app["application"]["application_name"],
-            appName: app["application"]["package_name"],
-            appPkgName: app["application"]["package_name"],
-            "appPkgName": appPkgName,
-            "packageName": app["application"]["package_name"],
-            "id": app["id"],
-            "app_state": None,
-        }
-    elif hasattr(app, "application_name"):
-        appName = app.application_name
-        appPkgName = appName + (" (%s)" % app.package_name)
-        entry = {
-            "app_name": app.application_name,
-            appName: app.package_name,
-            appPkgName: app.package_name,
-            "appPkgName": appPkgName,
-            "packageName": app.package_name,
-            "versions": app.versions,
-            "id": app.id,
-        }
-    else:
-        appName = app["app_name"]
-        appPkgName = appName + (" (%s)" % app["package_name"])
-        entry = {
-            "app_name": app["app_name"],
-            appName: app["package_name"],
-            appPkgName: app["package_name"],
-            "appPkgName": appPkgName,
-            "packageName": app["package_name"],
-            "app_state": app["state"],
-            "id": app["id"],
-        }
-
-    if Globals.frame and hasattr(Globals.frame, "sidePanel"):
-        selectedDeviceAppsMatch = list(
-            filter(
-                lambda entry: entry["app_name"] == appName
-                and entry["appPkgName"] == appPkgName,
-                Globals.frame.sidePanel.selectedDeviceApps,
-            )
-        )
-        enterpriseAppsMatch = list(
-            filter(
-                lambda entry: entry["app_name"] == appName
-                and entry["appPkgName"] == appPkgName,
-                Globals.frame.sidePanel.enterpriseApps,
-            )
-        )
-        if selectedDeviceAppsMatch:
-            indx = Globals.frame.sidePanel.selectedDeviceApps.index(
-                selectedDeviceAppsMatch[0]
-            )
-            oldEntry = Globals.frame.sidePanel.selectedDeviceApps[indx]
-            if update:
-                oldEntry.update(entry)
-                Globals.frame.sidePanel.selectedDeviceApps[indx] = entry = oldEntry
-        if enterpriseAppsMatch:
-            indx = Globals.frame.sidePanel.enterpriseApps.index(enterpriseAppsMatch[0])
-            oldEntry = Globals.frame.sidePanel.enterpriseApps[indx]
-            if update:
-                oldEntry.update(entry)
-                Globals.frame.sidePanel.enterpriseApps[indx] = entry = oldEntry
-
-    return entry
